Rename main resolvers file for clarity and organize resolvers into separate folders and files
