+++ src/resolvers/enums/gender.ts
@@ -1,0 +1,7 @@
+const genderEnumsResolvers = {
+  Male: "Male",
+  Female: "Female",
+  Other: "Other",
+};
+
+export default genderEnumsResolvers;
