+++ src/resolvers/enums/height.ts
@@ -1,0 +1,7 @@
+const heightEnumsResolvers = {
+  Tall: "Tall",
+  Medium: "Medium",
+  Short: "Short",
+};
+
+export default heightEnumsResolvers;
