+++ src/resolvers/enums/region.ts
@@ -1,0 +1,11 @@
+const regionEnumsResolvers = {
+  Mondstadt: "Mondstadt",
+  Liyue: "Liyue",
+  Inazuma: "Inazuma",
+  Sumeru: "Sumeru",
+  Fontaine: "Fontaine",
+  Natlan: "Natlan",
+  Snezhnaya: "Snezhnaya",
+};
+
+export default regionEnumsResolvers;
