+++ src/resolvers/enums/stat.ts
@@ -1,0 +1,20 @@
+const statEnumsResolvers = {
+  Anemo_DMG_Bonus: "Anemo DMG Bonus",
+  ATK: "ATK",
+  CRIT_DMG: "CRIT DMG",
+  CRIT_Rate: "CRIT Rate",
+  Cryo_DMG_Bonus: "Cryo DMG Bonus",
+  DEF: "DEF",
+  Dendro_DMG_Bonus: "Dendro DMG Bonus",
+  Electro_DMG_Bonus: "Electro DMG Bonus",
+  Elemental_Mastery: "Elemental Mastery",
+  Energy_Recharge: "Energy Recharge",
+  Geo_DMG_Bonus: "Geo DMG Bonus",
+  Healing_Bonus: "Healing Bonus",
+  HP: "HP",
+  Hydro_DMG_Bonus: "Hydro DMG Bonus",
+  Physical_DMG_Bonus: "Physical DMG Bonus",
+  Pyro_DMG_Bonus: "Pyro DMG Bonus",
+};
+
+export default statEnumsResolvers;
